+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.0.4 (2017-12-15)
+- Added "Code Intelligence Status" badge.
+- Added Registry class.
+- Added unitTest for class Brownie\Util\Registry.
+
 ## 0.0.3 (2017-12-14)
 - Fixed Scrutinizer issues.